created duck for investmentsReducer
