+++ package.json
@@ -3,6 +3,7 @@
   "private": true,
   "dependencies": {
     "react-redux": "^5.0.7",
-    "redux": "^4.0.0"
+    "redux": "^4.0.0",
+
   }
 }
