+++ user-privileges/package.json
@@ -7,6 +7,7 @@
     "humanize-plus": "^1.7.0",
     "isomorphic-fetch": "^2.2.1",
     "jest-fetch-mock": "^1.6.5",
+    "prop-types": "^15.6.2",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
     "react-pure-lifecycle": "^3.0.0",
