+++ user-privileges/src/components/AvailableOfferings.js
@@ -1,4 +1,5 @@
 import {connect} from 'react-redux';
+import {addNewInvestment} from '../redux/modules/InvestmentsReducer';
 import Action from './Actions';
 import Client from './Client';
 import OfferingCards from './OfferingCards';
@@ -13,7 +14,7 @@ const mapDispatchToOfferingCardsProps = (dispatch, props) =>(
     )
   ),
   onInvestmentFormSubmit: (investment) =>(
-    dispatch(Action.addNewInvestment(investment))
+    dispatch(addNewInvestment(investment))
   )
 }
 );
