+++ user-privileges/src/components/InvestorDashboard.js
@@ -1,5 +1,5 @@
 import {connect} from 'react-redux';
-import Action from './Actions';
+import {updateInvestment,setInvestments} from '../redux/modules/InvestmentsReducer';
 import Client from './Client';
 import InvestmentList from './InvestmentList';
 
@@ -9,12 +9,12 @@ const mapDispatchToInvestmentListProps = (dispatch,props)=>(
     onMount:() =>Client.getInvestments(
       (investments)=>(
         dispatch(
-          Action.setInvestments(investments)
+          setInvestments(investments)
         )
       )
     ),
     editInvestment:(investment) => dispatch(
-      Action.updateInvestment(investment)
+      updateInvestment(investment)
     ),
 
   }
