+++ user-privileges/src/redux/modules/CurrentUserReducer.js
