+++ user-privileges/src/redux/modules/OfferingsReducer.js
